Create Transform class
